Rename and remove some props
