+++ Arcade.sln
@@ -28,7 +28,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Arcade.Sdk
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Helix.JobSender", "src\Microsoft.DotNet.Helix\JobSender\Microsoft.DotNet.Helix.JobSender.csproj", "{D6FFE1CC-98D6-425F-95D2-E40122D96E80}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Xunit.NetCore.Extensions", "src\xunit.netcore.extensions\src\Xunit.NetCore.Extensions.csproj", "{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.XUnitExtensions", "src\Microsoft.DotNet.XUnitExtensions\src\Microsoft.DotNet.XUnitExtensions.csproj", "{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -160,18 +160,18 @@ Global
 		{D6FFE1CC-98D6-425F-95D2-E40122D96E80}.Release|x64.Build.0 = Release|Any CPU
 		{D6FFE1CC-98D6-425F-95D2-E40122D96E80}.Release|x86.ActiveCfg = Release|Any CPU
 		{D6FFE1CC-98D6-425F-95D2-E40122D96E80}.Release|x86.Build.0 = Release|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Debug|x64.Build.0 = Debug|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Debug|x86.Build.0 = Debug|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Release|x64.ActiveCfg = Release|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Release|x64.Build.0 = Release|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Release|x86.ActiveCfg = Release|Any CPU
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A}.Release|x86.Build.0 = Release|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Debug|x64.Build.0 = Debug|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Debug|x86.Build.0 = Debug|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Release|x64.ActiveCfg = Release|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Release|x64.Build.0 = Release|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Release|x86.ActiveCfg = Release|Any CPU
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -188,7 +188,7 @@ Global
 		{F7E1C1EF-C234-41BA-A24F-FAD5F357562C} = {730AF68A-22FD-4372-9341-5F67D7701DA7}
 		{747A5C75-6069-4C45-8049-AEAA1D864105} = {730AF68A-22FD-4372-9341-5F67D7701DA7}
 		{D6FFE1CC-98D6-425F-95D2-E40122D96E80} = {6BE49638-F842-4329-BE8A-30C0F30CCAA5}
-		{BE8ED8C1-C314-4C4E-A929-64C9C8B3552A} = {730AF68A-22FD-4372-9341-5F67D7701DA7}
+		{F8E7FEC3-B9E0-4A08-8608-F48379204F3A} = {730AF68A-22FD-4372-9341-5F67D7701DA7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {06E6CFF5-C54E-4623-9587-71A9E3F041D9}
