+++ src/Microsoft.DotNet.XUnitExtensions/src/Discoverers/PlatformSpecificDiscoverer.cs
@@ -6,8 +6,9 @@ using System.Linq;
 using System.Collections.Generic;
 using Xunit.Abstractions;
 using Xunit.Sdk;
+using Xunit;
 
-namespace Xunit.NetCore.Extensions
+namespace Microsoft.DotNet.XUnitExtensions
 {
     /// <summary>
     /// This class discovers all of the tests and test classes that have
