--- src/xunit.netcore.extensions/README.md
@@ -1,1 +1,0 @@
-# Xunit.NetCore.Extensions
