--- src/xunit.netcore.extensions/src/Xunit.NetCore.Extensions.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <OutputType>Library</OutputType>
-    <AssemblyName>xunit.netcore.extensions</AssemblyName>
-    <IsPackable>true</IsPackable>
-    
-    <PackageId>Microsoft.xunit.netcore.extensions</PackageId>
-    <Authors>Microsoft</Authors>
-    <VersionPrefix>2.4.0</VersionPrefix>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <Description>This package provides things like various traits and discovers like OuterLoop/ActiveIssue that are used by .NET Core framework test projects.</Description>
-    <Copyright>Copyright © Microsoft Corporation</Copyright>
-    <PackageLicenseUrl>http://go.microsoft.com/fwlink/?LinkId=329770</PackageLicenseUrl>
-    <PackageProjectUrl>https://github.com/dotnet/arcade</PackageProjectUrl>
-    <PackageIconUrl>http://go.microsoft.com/fwlink/?LinkID=288859</PackageIconUrl>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="xunit" Version="$(XUnitVersion)" />
-  </ItemGroup>
-
-</Project>
